+++ tests/integration_tests/test_client.py
@@ -216,7 +216,10 @@ def test_empty_result(test_client: Client):
     assert len(test_client.query("SELECT * FROM system.tables WHERE name = '_NOT_A THING'").result_rows) == 0
 
 
-def test_temporary_tables(test_client: Client):
+def test_temporary_tables(test_client: Client, test_config: TestConfig):
+    if test_config.cloud:
+        pytest.skip("Skipping temporary tables test in cloud env")
+
     session_id = test_client.get_client_setting("session_id")
     session_settings = {"session_id": session_id}
     test_client.command("""
@@ -235,7 +238,7 @@ def test_temporary_tables(test_client: Client):
     test_client.insert_df('temp_test_table', df, settings=session_settings)
     df = test_client.query_df('SELECT * FROM temp_test_table', settings=session_settings)
     assert len(df['field1']) == 4
-    test_client.command('DROP TABLE temp_test_table', settings=session_settings)
+    test_client.command('DROP TABLE IF EXISTS temp_test_table', settings=session_settings)
 
 
 def test_str_as_bytes(test_client: Client, table_context: Callable):